Fix scaling to preserve page numbers of pdf file
Fix the %%scale directives to preserve the legacy page
numbers when generating the output ps and pdf files.

The scaling directives in the original .abc file did not
cause the output .ps/.pdf files to have the "correct"
page numbers.  There were many instances where a tune
would not fit on the "correct" page, and would move to
the next page.

The default scaling is 0.75, but several pages require
a scaling of 0.7 or even 0.65 in order to fit all its
tunes on the original page.

When possible, pages use a scaling factor of 0.8, as tunes
look better when made slightly bigger.